+++ src/main/java/seedu/address/logic/parser/SetCommandParser.java
@@ -43,11 +43,12 @@ public class SetCommandParser implements Parser<SetCommand> {
         }
 
         /* handles overflow value */
-        if (Integer.parseInt(argMultimap.getValue(PREFIX_AMOUNT).get()) >= 1000000) {
+        if (Double.parseDouble(argMultimap.getValue(PREFIX_AMOUNT).get()) >= 1000000) {
             throw new ParseException(String.format(SetCommand.MESSAGE_AMOUNT_OVERFLOW));
         }
 
         Amount budget = ParserUtil.parseAmount(argMultimap.getValue(PREFIX_AMOUNT).get());
+
         Date date = ParserUtil.parseDate(argMultimap.getValue(PREFIX_DATE).get());
 
         if (date.isPast()) {
